requirements.txt
argh===0.26.2
psycopg2===2.8.5
python-telegram-bot===12.8
pyyaml===5.3.1
watchdog===0.10.3
yoyo-migrations===7.2.0
